mailbox plugin failed to install django_mailbox

+++ lino_xl/lib/mailbox/models.py
@@ -1,24 +1,16 @@
 # -*- coding: UTF-8 -*-
+# Copyright 2013-2021 Rumma & Ko Ltd
+# License: BSD (see file COPYING for details)
 
-"""Database models for `lino_xl.lib.mailbox`.
-
-"""
-import logging ; logger = logging.getLogger(__name__)
+from django.db import models
 
 try:
-    from django_mailbox import models
+    import django_mailbox
 except ImportError:
-    pass
+    django_mailbox = None
 
-    # without django_mailbox  this plugin won't work, but it must be importable
-    # for :manage:`install`.
+from lino.api import dd, rt, _
 
-from django.utils.translation import gettext_lazy as _
-
-from lino.api import dd, rt
-
-#
-#
 
 def preview(obj, ar):
     return obj.html or obj.text
@@ -31,50 +23,54 @@ def spam(obj):
     else:
         return False
 
+if django_mailbox is not None:
+    # When django_mailbox is not installed, this plugin does nothing, but it
+    # must be importable for :manage:`install`.
 
-dd.inject_field('django_mailbox.Message', 'preview',
-                dd.VirtualField(dd.HtmlBox(_("Preview")), preview))
-dd.inject_field('django_mailbox.Message', 'ticket',
-                dd.ForeignKey('tickets.Ticket', blank=True, null=True))
+    dd.inject_field('django_mailbox.Message', 'preview',
+                    dd.VirtualField(dd.HtmlBox(_("Preview")), preview))
+    dd.inject_field('django_mailbox.Message', 'ticket',
+                    dd.ForeignKey('tickets.Ticket', blank=True, null=True))
+    # dd.inject_field('django_mailbox.Message', 'spam',
+    #                 models.BooleanField(_("Spam"), default=False))
+    #
+    dd.update_field('django_mailbox.Message', 'from_header', format="plain")
 
-dd.update_field('django_mailbox.Message', 'from_header', format="plain")
-
-
-from .ui import *
-
-@dd.schedule_often(10)
-def get_new_mail():
-    for mb in rt.models.django_mailbox.Mailbox.objects.filter(active=True):
-        mails = mb.get_new_mail()
-        for mail in mails:
-            if spam(mail):
-                mail.spam = True
-                mail.full_clean()
-                mail.save()
-        if mails:
-            logger.info("got {} from mailbox: {}".format(mails,mb))
-
-
-class DeleteSpam(dd.Action):
-
-    show_in_bbar = True
-    readonly = False
-    # required_roles = dd.login_required(Worker)
-    label = u"X"
+    @dd.schedule_often(10)
+    def get_new_mail():
+        for mb in rt.models.django_mailbox.Mailbox.objects.filter(active=True):
+            # print("20210305", mb)
+            mails = mb.get_new_mail()
+            for mail in mails:
+                if spam(mail):
+                    mail.spam = True
+                    mail.full_clean()
+                    mail.save()
+            if mails:
+                dd.logger.info("got {} from mailbox: {}".format(mails, mb))
 
 
-    def run_from_ui(self, ar, **kw):
-        spams = rt.models.django_mailbox.Message.objects.filter(spam = True)
-        logger.info("Deleting spam messages [%s]", spams)
+    class DeleteSpam(dd.Action):
 
-        def ok(ar):
-            for obj in spams:
-                obj.delete()
-            ar.set_response(refresh=True)
+        show_in_bbar = True
+        readonly = False
+        # required_roles = dd.login_required(Worker)
+        label = "X"
 
-        ar.confirm(
-            ok,
-            _("Delete {} messages.").format(spams.count()),
-            _("Are you sure?"))
+        def run_from_ui(self, ar, **kw):
+            spams = rt.models.django_mailbox.Message.objects.filter(spam=True)
+            dd.logger.info("Deleting spam messages [%s]", spams)
 
-dd.inject_action("django_mailbox.Message", delete_spam=DeleteSpam())
+            def ok(ar):
+                for obj in spams:
+                    obj.delete()
+                ar.set_response(refresh=True)
+
+            ar.confirm(
+                ok,
+                _("Delete {} messages.").format(spams.count()),
+                _("Are you sure?"))
+
+    dd.inject_action("django_mailbox.Message", delete_spam=DeleteSpam())
+
+    from .ui import *
